backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
# MongoDB async driver
motor>=3.3.2
pymongo>=4.6.1
# Environment variables
python-dotenv>=1.0.0
# Gemini
google-generativeai>=0.8.0
# Eleven Labs
elevenlabs>=1.0.0
# CORS (included with fastapi, listed explicitly for clarity)
httpx>=0.27.0
# PDF text extraction (optional – for research PDFs)
PyPDF2>=3.0.0
